style: move deduplicate
